requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0
undetected_chromedriver